Extract QuestExceptionHandler from Objective
